Fix merging of edit_roi (cloud version different from barebone, and snakemake9.6 different from barebone)
